requirements.txt
Django==4.2.1
python-environ==0.4.54
djangorestframework==3.14.0
djoser==2.2.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
drf-yasg==1.21.5
celery==5.2.7
redis==4.5.5
PyYAML==6.0
pytest==7.3.1